Add transpose, determinant, minor and submatrix functions for matrices
