+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/02/04 21:14:15 by coder             #+#    #+#              #
-#    Updated: 2022/10/07 16:22:35 by sguilher         ###   ########.fr        #
+#    Updated: 2022/10/07 17:54:20 by sguilher         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -24,7 +24,7 @@ SRCS =			main.c handle_input.c mlx_utils.c colors.c utils.c ray.c \
 				sphere.c $(TUPLES) $(MATRIX)
 TUPLES =		point.c vector.c vector_and_point_operations.c \
 				vector_operations.c
-MATRICES =		matrix.c matrix_operations.c
+MATRICES =		matrix.c matrix_multiply.c matrix_operations.c
 OBJS_DIR =		./obj
 OBJS	=		$(SRCS:%.c=$(OBJS_DIR)/%.o)
 HEADERS	=		minirt.h
