+++ src/matrix/matrix_operations.c
@@ -6,66 +6,103 @@
 /*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 14:21:54 by elraira-          #+#    #+#             */
-/*   Updated: 2022/10/07 17:39:11 by sguilher         ###   ########.fr       */
+/*   Updated: 2022/10/07 18:46:37 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "minirt.h"
+#include "minirt.h"
 
-t_matrix multiply_matrix(t_matrix m1, t_matrix m2)
+/**
+ * @brief Matrix transposition is a process of interchanging rows and columns
+ * of a matrix. It will be used when translating normal vectors between objects
+ * and the space. It is crucial when shading objects.
+ *
+ * @param matrix The matrix to be transposed.
+ * @return t_matrix The transposed matrix.
+ */
+t_matrix	matrix_transpose(t_matrix matrix)
 {
-	t_matrix res;
+	t_matrix	res;
+	int			i;
+	int			j;
+
+	res = create_matrix(matrix.size);
+	i = 0;
+	while (i < matrix.size)
+	{
+		j = 0;
+		while (j < matrix.size)
+		{
+			res.matrix[i][j] = matrix.matrix[j][i];
+			j++;
+		}
+		i++;
+	}
+	return (res);
+}
+
+t_matrix	get_submatrix(t_matrix m, int column)
+{
+	t_matrix submatrix;
 	int i;
 	int j;
 	int k;
 
-	if (m1.size != m2.size)
+	submatrix = create_matrix(m.size - 1);
+	i = 1;
+	while (i < submatrix.size)
 	{
-		printf("Error: matrix size mismatch\n");
-		return (m1);
-	}
-	i = -1;
-	res = create_matrix(m1.size);
-	while(++i < m1.size)
-	{
-		j = -1;
-		while (++j < m2.size)
+		j = 0;
+		k = 0;
+		while (j < submatrix.size)
 		{
-			k = -1;
-			res.matrix[i][j] = 0;
-			while (++k < m1.size)
-				res.matrix[i][j] += m1.matrix[i][k] * m2.matrix[k][j];
+			if (k != column)
+			{
+				submatrix.matrix[i - 1][j] = m.matrix[i][k];
+				j++;
+			}
+			k++;
+		}
+		i++;
+	}
+	return (submatrix);
+}
+
+double	get_matrix_minor(t_matrix m, int column)
+{
+	t_matrix submatrix;
+
+	submatrix = get_submatrix(m, column);
+	if (column / 2 == 0)
+		return (get_matrix_determinant(submatrix));
+	else
+		return (-get_matrix_determinant(submatrix));
+}
+
+double	get_matrix_determinant(t_matrix m)
+{
+	double	det;
+	int		i;
+
+	if (m.size == 2)
+		det = m.matrix[0][0] * m.matrix[1][1]
+			- m.matrix[0][1] * m.matrix[1][0];
+	else
+	{
+		det = 0;
+		i = 0;
+		while (i < m.size)
+		{
+			det += m.matrix[0][i] * get_matrix_minor(m, i);
+			i++;
 		}
 	}
-	return (res);
+	return (det);
 }
 
-t_vector	multiply_matrix_by_vector(t_matrix m, t_vector v)
-{
-	t_vector res;
-
-	res.x = m.matrix[0][0] * v.x + m.matrix[0][1] * v.y
-		+ m.matrix[0][2] * v.z + m.matrix[0][3] * v.w;
-	res.y = m.matrix[1][0] * v.x + m.matrix[1][1] * v.y
-		+ m.matrix[1][2] * v.z + m.matrix[1][3] * v.w;
-	res.z = m.matrix[2][0] * v.x + m.matrix[2][1] * v.y
-		+ m.matrix[2][2] * v.z + m.matrix[2][3] * v.w;
-	res.w = m.matrix[3][0] * v.x + m.matrix[3][1] * v.y
-		+ m.matrix[3][2] * v.z + m.matrix[3][3] * v.w;
-	return (res);
-}
-
-t_point	multiply_matrix_by_point(t_matrix m, t_point p)
-{
-	t_point res;
-
-	res.x = m.matrix[0][0] * p.x + m.matrix[0][1] * p.y
-		+ m.matrix[0][2] * p.z + m.matrix[0][3] * p.w;
-	res.y = m.matrix[1][0] * p.x + m.matrix[1][1] * p.y
-		+ m.matrix[1][2] * p.z + m.matrix[1][3] * p.w;
-	res.z = m.matrix[2][0] * p.x + m.matrix[2][1] * p.y
-		+ m.matrix[2][2] * p.z + m.matrix[2][3] * p.w;
-	res.w = m.matrix[3][0] * p.x + m.matrix[3][1] * p.y
-		+ m.matrix[3][2] * p.z + m.matrix[3][3] * p.w;
-	return (res);
-}
+/*
+A submatrix is what is left when you delete a single row and column from
+a matrix.
+The minor of an element at row i and column j is the determinant of the
+sub- matrix at (i,j).
+Cofactors are are minors that have (possibly) had their sign changed.*/
